+++ app/apps.py
@@ -6,4 +6,4 @@ class AppConfig(AppConfig):
     name = 'app'
 
     def ready(self):
-        import app.signals
+        import app.signals.election_signals
