+++ app/signals/election_signals.py
